Move `json_utils.py` to `autogpt/core`
