+++ autogpts/autogpt/autogpt/agent_factory/profile_generator.py
@@ -3,7 +3,7 @@ import logging
 
 from autogpt.config import AIDirectives, AIProfile, Config
 from autogpt.core.configuration import SystemConfiguration, UserConfigurable
-from autogpt.core.json_utils.utils import json_loads
+from autogpt.core.json_utils import json_loads
 from autogpt.core.prompting import (
     ChatPrompt,
     LanguageModelClassification,
