--- autogpts/autogpt/autogpt/core/json_utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utility functions for working with JSON files."""
