+++ autogpts/autogpt/autogpt/core/planning/prompt_strategies/initial_plan.py
@@ -1,7 +1,7 @@
 import logging
 
 from autogpt.core.configuration import SystemConfiguration, UserConfigurable
-from autogpt.core.json_utils.utils import json_loads
+from autogpt.core.json_utils import json_loads
 from autogpt.core.planning.schema import Task, TaskType
 from autogpt.core.prompting import PromptStrategy
 from autogpt.core.prompting.schema import ChatPrompt, LanguageModelClassification
