+++ autogpts/autogpt/autogpt/processing/text.py
@@ -6,7 +6,7 @@ from typing import Iterator, Optional, TypeVar
 import spacy
 
 from autogpt.config import Config
-from autogpt.core.json_utils.utils import extract_list_from_json
+from autogpt.core.json_utils import extract_list_from_json
 from autogpt.core.prompting import ChatPrompt
 from autogpt.core.resource.model_providers import (
     ChatMessage,
